began working on tests
